% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{currentseason_helper}
\alias{currentseason_helper}
\title{Helper function}
\usage{
currentseason_helper(player)
}
\arguments{
\item{player}{individual player (list) returned by \link{collect_ff}}
}
\description{
currentseason_helper is called by \link{currentseason_df} to
collate data for the current season
}

