% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasysocceR.R
\docType{package}
\name{fantasysocceR}
\alias{fantasysocceR}
\alias{fantasysocceR-package}
\title{fantasysocceR}
\description{
Datasets for players and teams in the fantasy football (soccer)
game (fantasy.premierleague.com).  The game data updates frequently as users
'manage' their teams, so datasets can go out of date quickly.  I will try
to update datasets regularly, but functions are available to help collect
up-to-date data in the same/similar format to the existing, stable, datasets.

Datasets include:
\itemize{
     \item \strong{players} Player summary data for 2015/16 season, see \link{players}
     \item \strong{pastseasons} Past season data for the 2015/16 season, see
     \link{pastseasons}
     \item \strong{players201415} Player summary data for 2014/15 season, 
     see \link{players201415}
     \item \strong{pastseasons201415} Past season data for 2014/15 season, 
     see \link{pastseasons201415}
     \item \strong{season201415} Performance data for the 2014/15 Season, see \link{season201415}
}
}
\details{
Fantasy Football (soccer) datasets and functions
}

