% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fantasysocceR.R
\docType{data}
\name{gameweeks}
\alias{gameweeks}
\title{Gameweeks dataset}
\format{760 rows, 5 columns}
\description{
Data about the gameweeks, which team will face which team.  Like
the \link{players} dataset, this will likely change as/when teams are forced
to change their fixture schedule.  See \link{collect_ff} and related
functions for getting up-to-date data.
}
\details{
Variables
\itemize{
     \item \strong{date} fixture data
     \item \strong{gw} gameweek
     \item \strong{team} team
     \item \strong{home} are team at home (\code{TRUE}), or away
     (\code{FALSE})
     \item \strong{opp} opponent
}
}
\keyword{datasets}

