% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameweeks_df.R
\name{gameweeks_df}
\alias{gameweeks_df}
\title{Construct gameweeks dataframe}
\usage{
gameweeks_df(player_list)
}
\arguments{
\item{\strong{player_list}}{list returned by \link{collect_ff}}
}
\description{
Converts the list returned by \link{collect_ff} into a dataframe
of all the gameweeks for each of the 20 sides in the Premiership.  The 5
variables are:
\itemize{
     \item \strong{team} team
     \item \strong{date} fixture data
     \item \strong{gw} gameweek
     \item \strong{opp} opponent
     \item \strong{home} are \strong{team} at home (\code{TRUE}), or away
     (\code{FALSE})
}
}
\examples{
## Collect fresh data
player_list <- collect_ff(n = 10)
gameweeks <- gameweeks_df(player_list)
}

