% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fantasysocceR.R
\docType{data}
\name{pastseasons}
\alias{pastseasons}
\title{Past seasons dataset}
\format{1920 rows, 24 columns}
\description{
Data about the current players in the game and their performance
in past seasons.  Unlike \link{players} dataset, the
majority of the variables in this dataset should be stable, but it does
include some of the data found in \link{players}, data collected Jul 15, 2015.
}
\details{
If a player was included in more than one previous version of the
game, then that player will have multiple rows.  If a player is new to the
game, they will have summary data found in \link{players} but past seasons
will be filled with NAs.
}
\keyword{datasets}

