% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pastseasons_df}
\alias{pastseasons_df}
\title{Construct past seasons dataframe from fresh fantasy data}
\usage{
pastseasons_df(player_list, add_vars = NULL)
}
\arguments{
\item{\strong{player_list}}{list of players (the list returned by
\link{collect_ff})}
}
\description{
Converts the list returned by \link{collect_df} into a dataframe
with past performances of all current players in the game.
}
\details{
This function calls the \link{players_df} function to provide
summary data for the current season.  The dataframe returned consists of:
\itemize{
     \item \strong{pos} player pos
     \item \strong{team} player team
     \item \strong{pts} total points scored last season
     \item \strong{id} player id
     \item \strong{name} player name
     \item \strong{value} player value
     \item \strong{pct} percentage of fantasy teams player is selected by
     \item \strong{season} previous season
     \item \strong{mins} number of minutes played
     \item \strong{goals} goals scored
     \item \strong{assists} assists
     \item \strong{cs} clean sheets
     \item \strong{og} own goals
     \item \strong{ga} goals against
     \item \strong{pens_svd} penalties saved
     \item \strong{pens_msd} penalties missed
     \item \strong{yel} yellow cards
     \item \strong{red} red cards
     \item \strong{saves} saves
     \item \strong{bonus} bonus points
     \item \strong{ea_ppi} EA sports player performance index
     \item \strong{bps} bonus points system
     \item \strong{fin_val} value at end of the season
     \item \strong{ssn_pts} total points for the season
}
}
\examples{
## Collect fresh data
player_list <- collect_ff(n = 10)
pastseasons <- pastseasons_df(player_list)
}

