% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pastseasons_helper}
\alias{pastseasons_helper}
\title{Helper function}
\usage{
pastseasons_helper(player)
}
\arguments{
\item{player}{individual player (list) returned by \link{collect_ff}}
}
\description{
pastseasons_helper is called by \link{pastseasons_df} to 
collate data for players who
}

