% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fantasysocceR.R
\docType{data}
\name{season201415}
\alias{season201415}
\title{Season 2014/15 dataset}
\format{24876 rows, 27 columns}
\description{
dataset of fantasy football performance scores for each player
included in the game and each gameweek
}
\details{
Variables
\itemize{
     \item \strong{id} player id
     \item \strong{name} player name
     \item \strong{pos} player position
     \item \strong{team} player's team
     \item \strong{pts} end of season points total
     \item \strong{value} end of season player value
     \item \strong{pct} percentage of teams player is included in at end of season
     \item \strong{date} gameweek date
     \item \strong{gw} gameweek id
     \item \strong{opp} opponent and result
     \item \strong{mins} minutes played
     \item \strong{goals} goals scored
     \item \strong{assists} assists
     \item \strong{cs} clean sheets
     \item \strong{ga} goals against
     \item \strong{og} own goals
     \item \strong{pens_svd} penalties saved
     \item \strong{pens_mds} penalties missed
     \item \strong{yel} yellow cards
     \item \strong{red} red cards
     \item \strong{saves} saves
     \item \strong{bonus} bonus points
     \item \strong{ea_ppi} ea sports player performance index
     \item \strong{bps} bonus points system
     \item \strong{net_transfers} transfers in/out of fantasy teams
     \item \strong{gw_val} gameweek value
     \item \strong{gw_pts} gameweek points
}
}
\keyword{datasets}

