% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/team_pal.R
\name{team_pal}
\alias{team_pal}
\title{Premiership Team colors palette}
\usage{
team_pal(teams = NULL)
}
\arguments{
\item{\strong{teams}}{specify teams (not always necessary)}
}
\description{
Palette of primary team colors where possible - the secondary color is used
when the primary color is white.
}
\details{
It returns a named vector, with the team names matching the format
of the datasets in this package.  It can simply be called within a ggplot
plot, see \emph{examples}
}
\examples{
team_pal(teams = c("Arsenal", "Aston Villa"))
   Arsenal Aston Villa
 "#F12933"   "#A32249"

# using within ggplot
data(players)
ggplot(players, aes(x=value, y=pts, color=team)) +
     geom_point() +
     scale_color_manual(values = team_pal())
}

