% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_pal.R
\docType{data}
\name{epl_palettes}
\alias{epl_palettes}
\title{List of Premiership team palettes}
\format{An object of class \code{list} of length 23.}
\usage{
epl_palettes
}
\description{
each team, 23 total teams, from seasons 2014/15 and 2015/16 has
three colours
}
\details{
Use \code{\link{team_pal}} to construct palettes
}
\keyword{datasets}

