% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fantasysocceR}
\alias{fantasysocceR}
\alias{fantasysocceR-package}
\title{fantasysocceR}
\description{
Datasets for players and teams in the fantasy football (soccer)
game (fantasy.premierleague.com).  The game data updates frequently as users
'manage' their teams, so datasets can go out of date quickly.  I will try
to update datasets regularly, but functions are available to help collect
up-to-date data in the same/similar format to the existing, stable, datasets.

Datasets include:
\itemize{
     \item \strong{players} Player summary data, see \link{players}
     \item \strong{pastseasons} Past season data, see \link{pastseasons}
     \item \strong{gameweeks} Fixture data, see \link{gameweeks}
}
}
\details{
Fantasy Football (soccer) datasets and functions
}

