% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasysocceR.R
\docType{data}
\name{players}
\alias{players}
\title{Players dataset}
\format{532 rows, 7 columns}
\description{
Summary data about the current players in the game, some 
variables in this dataset will almost immediately be out of date (such as 
\emph{pct}), data collected Aug 02, 2015.
See \link{collect_ff} for getting up-to-date data.
}
\details{
Variables
\itemize{
     \item \strong{pos} player pos
     \item \strong{team} player team
     \item \strong{pts} total points scored last season
     \item \strong{id} player id
     \item \strong{name} player name
     \item \strong{value} player value
     \item \strong{pct} percentage of fantasy teams player is selected by
}
}
\keyword{datasets}

