% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{players_helper}
\alias{players_helper}
\title{Helper function}
\usage{
players_helper(player, add_vars = NULL)
}
\arguments{
\item{player}{individual player (list) returned by \link{collect_ff}}

\item{add_vars}{any additional variables to return from list}
}
\description{
players_helper is called by both \link{players_df} and 
\link{pastseasons_df} to collate summary data for each player.
}

