% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_pal.R
\name{team_pal}
\alias{team_pal}
\title{Premiership Team colors palette}
\usage{
team_pal(teams = NULL, simple = TRUE, n = 1)
}
\arguments{
\item{teams}{specify teams (not always necessary)}

\item{simple}{return a single colour per team}

\item{n}{number of colours to return in a palette (per team) (maximum of 3)}
}
\description{
Palette of primary team colors where possible - the secondary color is used
when the primary color is white.
}
\details{
It returns a named vector, with the team names matching the format 
of the datasets in this package.  It can simply be called within a ggplot 
plot, see \emph{examples}
}
\examples{
\dontrun{
team_pal(teams = c("Arsenal", "Aston Villa"))

# using within ggplot
data(players)
ggplot(players, aes(x=value, y=pts, color=team)) +
     geom_point() +
     scale_color_manual(values = team_pal()) 
}
}

